✨ Bimillah final(kureng notfikasi)

+++ src/services/kegiatanService.ts
@@ -1,4 +1,5 @@
 import * as kegiatanModels from '../models/kegiatanModels'
+import * as usersModels from '../models/usersModels'
 import { ChatRoom, Message } from '../models/livechatModels'
 import { addTimestamps } from '../models/utilsModel'
 
@@ -6,17 +7,16 @@ export async function fetchAllKegiatan() {
     return await kegiatanModels.fetchAllKegiatan()
 }
 
-export async function fetchKegiatanByUser(uidUser: string, status?: 'selesai' | 'ditugaskan', tanggal?: string,) {
-    const temp = await kegiatanModels.fetchKegiatanByUser(uidUser, status, tanggal, undefined)
-    if (!temp.userId) return "user_is_not_found"
-    if (temp.kegiatan.length === 0) return "kegiatan_is_not_found"
+export async function fetchKegiatanByUser(uidUser: string, isDone?: boolean, tanggal?: string,) {
+    const temp = await kegiatanModels.fetchKegiatanByUser(uidUser, isDone, tanggal, undefined)
+    if (temp.length === 0) return "kegiatan_is_not_found"
 
     return temp
 }
 
 export async function fetchKegiatan(uidKegiatan: string) {
     const temp = await kegiatanModels.fetchKegiatanByUid(uidKegiatan)
-    if (!temp) return "kegiatan_is_not_found"
+    if (!temp || Object.keys(temp!).length === 0) return "kegiatan_is_not_found"
 
     return temp
 }
@@ -36,18 +36,38 @@ export async function createKegiatan(kegiatanData: kegiatanModels.KegiatanDataTy
 
 export async function updateKegiatan(uidKegiatan: string, dataKegiatan: Partial<kegiatanModels.KegiatanDataType>, listKompetensiUid: string[]) {
     const temp = await kegiatanModels.fetchKegiatanOnly(uidKegiatan)
-    if (!temp) return "kegiatan_is_not_found"
+    if (!temp || Object.keys(temp).length === 0) return "kegiatan_is_not_found"
+
+    if (dataKegiatan.isDone && dataKegiatan.isDone == true) {
+        const listUser = await kegiatanModels.fetchKegiatanByUid(uidKegiatan)
+        if (listUser?.users && listUser.users.length != 0) {
+            const kegKomp = await kegiatanModels.fetchKompetensiKegiatan(uidKegiatan)
+            await Promise.all(listUser!.users.map(async (it) => {
+                await usersModels.addUserKompetensi(it.userId, kegKomp.kompetensi!)
+                await usersModels.addJumlahKegiatan(it.userId, undefined, kegKomp.tanggalMulai!.getFullYear(), kegKomp.tanggalMulai!.getMonth() + 1)
+
+            }))
+        }
+
+    }
 
     return await kegiatanModels.updateKegiatan(uidKegiatan, dataKegiatan, listKompetensiUid)
 }
 
 export async function deleteKegiatan(uidKegiatan: string) {
     const ap = await kegiatanModels.deleteKegiatan(uidKegiatan)
-    if (!ap) return "kegiatan_is_not_found"
+    if (!ap || Object.keys(ap).length === 0) return "kegiatan_is_not_found"
 
     // Rempve chatroom
     await ChatRoom.findOneAndDelete({ roomId: uidKegiatan })
     await Message.deleteMany({ roomId: uidKegiatan })
+
+    return ap
+}
+
+export async function deleteKompetensiKegiatan(uidKegiatan: string, uidKompetensi: string) {
+    const ap = await kegiatanModels.deleteKommpetensiFromkegiatan(uidKegiatan, uidKompetensi)
+    if (!ap || Object.keys(ap).length === 0) return "kegiatan_is_not_found"
 
     return ap
 }
